requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
idna==3.10
Jinja2==3.1.3
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
openai-whisper==20240930
pillow==10.2.0
psutil==6.1.0
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
regex==2024.11.6
requests==2.32.3
setuptools==70.0.0
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.0
typing_extensions==4.9.0
urllib3==2.2.3
